requirements.txt
aniso8601==1.3.0
astroid==1.6.0
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
flask-restful-swagger-2==0.35
flask-swagger-ui==3.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
PyJWT==1.5.3
pylint==1.8.1
PyMySQL==0.8.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
SQLAlchemy==1.2.0
Werkzeug==0.14.1
wrapt==1.10.11